[834] Follow Hudi's hudi-common package reorganization
Hudi master moved two packages that XTable depends on, in the
refactor(common) commits apache/hudi#19193 and apache/hudi#19195:

- org.apache.hudi.stats -> org.apache.hudi.metadata.stats, covering
  ValueMetadata, ValueType and HoodieColumnRangeMetadata. XTable's
  XTableValueMetadata and its test live inside Hudi's package to reach
  package-private members, so the files move directories as well.
- org.apache.hudi.avro.HoodieAvroUtils -> org.apache.hudi.common.avro.
  HoodieAvroUtils. The generated org.apache.hudi.avro.model classes did
  not move.

This is required to build against a Hudi release that carries the
reorganization, alongside the hudi.version bump this branch already
makes.
