+++ xtable-core/src/main/java/org/apache/hudi/metadata/stats/XTableValueMetadata.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
  
-package org.apache.hudi.stats;
+package org.apache.hudi.metadata.stats;
 
 import static org.apache.xtable.model.schema.InternalSchema.MetadataKey.TIMESTAMP_PRECISION;
 import static org.apache.xtable.model.schema.InternalSchema.MetadataValue.MICROS;
