+++ xtable-core/src/test/java/org/apache/hudi/metadata/stats/TestXTableValueMetadata.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
  
-package org.apache.hudi.stats;
+package org.apache.hudi.metadata.stats;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
